changing token to chat template
