+++ exo/inference/pytorch/inference.py
@@ -42,7 +42,14 @@ class PyTorchDynamicShardInferenceEngine(InferenceEngine):
         await self.ensure_shard(shard)
 
         inference_state = json.loads(inference_state) if inference_state else ""
-        tokens = self.tokenizer(prompt, return_tensors="pt").input_ids.to(self.device)
+        tokens = self.tokenizer.apply_chat_template(
+            conversation=[{
+                "role": "user", 
+                "content": prompt
+            }],
+            tokenize=True,
+            add_generation_prompt=False,
+        )
 
         if DEBUG >= 2:
             print(f"tokens: {tokens}\n")
